requirements.txt
# local package
# -e .

# external requirements
click==7.1.2
coverage
flake8
python-dotenv>=0.5.1
pytest
pyyaml==3.11
marshmallow-dataclass==8.3.0
dataclasses==0.6
pandas
numpy
scikit-learn==0.24.1
catboost

